% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid4-package.R
\docType{package}
\name{uuid4-package}
\alias{uuid4}
\alias{uuid4-package}
\title{uuid4: Generate Random UUID4 Strings}
\description{
Fast generation of random (i.e. version 4)
    universally unique identifiers (UUID4s) as specified in RFC 4122.
    Source of randomness is a PCG random number generator, separate from
    R's built-in RNGs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mikmart/uuid4}
  \item Report bugs at \url{https://github.com/mikmart/uuid4/issues}
}

}
\author{
\strong{Maintainer}: Mikko Marttila \email{mikmart@mikkomarttila.com}

Authors:
\itemize{
  \item Melissa O'Neill (PCG)
}

}
\keyword{internal}
