% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid.R
\name{uuid}
\alias{uuid}
\title{Generate a character vector of UUID4s}
\usage{
uuid(n = 1L)
}
\arguments{
\item{n}{number of UUIDs to generate}
}
\value{
a character vector of length \code{n}
}
\description{
Generate a character vector of UUID4s
}
\seealso{
\href{https://tools.ietf.org/html/rfc4122}{RFC 4122} for the specification
}
