% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid4-package.R
\docType{package}
\name{uuid4-package}
\alias{uuid4}
\alias{uuid4-package}
\title{uuid4: Generate UUID4s}
\description{
Fast generation of version 4 universally unique identifiers (UUID4s) as specified in [RFC 4122](https://tools.ietf.org/html/rfc4122).
}
\author{
\strong{Maintainer}: Mikko Marttila \email{mikmart@mikkomarttila.com}

}
\keyword{internal}
