% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid4-package.R
\docType{package}
\name{uuid4-package}
\alias{uuid4}
\alias{uuid4-package}
\title{uuid4: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
